turbovec-python/pyproject.toml
[build-system]
requires = ["maturin>=1.12,<2.0"]
build-backend = "maturin"

[project]
name = "turbovec"
version = "0.2.0"
description = "Fast vector quantization with 2-4 bit compression and SIMD search"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "Ryan Codrai" },
]
keywords = [
    "vector-search",
    "quantization",
    "nearest-neighbor",
    "ann",
    "simd",
    "rag",
    "embeddings",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Rust",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Database",
]

[project.urls]
Homepage = "https://github.com/RyanCodrai/turbovec"
Repository = "https://github.com/RyanCodrai/turbovec"
Issues = "https://github.com/RyanCodrai/turbovec/issues"

[project.optional-dependencies]
langchain = ["langchain-core>=0.3"]
llama-index = ["llama-index-core>=0.11"]
llm = ["numpy", "tiktoken"]
llm-openai = ["openai>=1.0"]
llm-anthropic = ["anthropic>=0.30"]
llm-gemini = ["google-genai>=0.3"]
llm-cohere = ["cohere>=5.0"]
llm-voyage = ["voyageai>=0.2"]
llm-local = ["sentence-transformers>=2.2"]
llm-redis = ["redis>=5.0"]
llm-proxy = ["fastapi>=0.110", "uvicorn>=0.27"]
llm-metrics = ["prometheus_client>=0.20"]

[tool.maturin]
manifest-path = "Cargo.toml"
features = ["pyo3/extension-module"]
python-source = "python"
module-name = "turbovec._turbovec"
